Refactor unit tests to check exception compatibility between Java MessageDigest and Digest
