+++ library/digest/src/commonTest/kotlin/org/kotlincrypto/core/DigestUnitTest.kt
@@ -17,66 +17,21 @@ package org.kotlincrypto.core
 
 import kotlin.test.Test
 import kotlin.test.assertEquals
-import kotlin.test.fail
 
-class DigestUnitTest {
+class DigestUnitTest: TestDigestException() {
 
     private val digest = TestDigest()
 
-    @Test
-    fun givenDigest_whenEmptyBytes_thenDoesNotThrow() {
-        digest.update(ByteArray(0))
+    override fun update(input: ByteArray) {
+        digest.update(input)
     }
 
-    @Test
-    fun givenDigest_whenLength0_thenThrowsExpected() {
-        // Input length check will pass here, but
-        // this would produce an ArrayIndexOutOfBounds exception
-        // unless len is 0
-        digest.update(ByteArray(0), -1, 0)
-
-        // This ensures that both jvm and non-Jvm check
-        // input length _before_ length == 0 which will
-        // simply return and not produce an error.
-        //
-        // If len > 0, this would throw an ArrayIndexOutOfBounds
-        // exception because offset > input.size
-        try {
-            digest.update(ByteArray(0), 1, 0)
-            fail()
-        } catch (_: IllegalArgumentException) {
-            // pass
-        }
+    override fun update(input: Byte) {
+        digest.update(input)
     }
 
-    @Test
-    fun givenDigest_whenLengthNegative_thenThrows() {
-        try {
-            digest.update(ByteArray(0), 0, -1)
-            fail()
-        } catch (_: IndexOutOfBoundsException) {
-            // pass
-        }
-    }
-
-    @Test
-    fun givenDigest_whenOffsetNegative_thenThrows() {
-        try {
-            digest.update(ByteArray(10), -1, 10)
-            fail()
-        } catch (_: IndexOutOfBoundsException) {
-            // pass
-        }
-    }
-
-    @Test
-    fun givenDigest_whenInputTooShort_thenThrows() {
-        try {
-            digest.update(ByteArray(5), -1, 10)
-            fail()
-        } catch (_: IllegalArgumentException) {
-            // pass
-        }
+    override fun update(input: ByteArray, offset: Int, len: Int) {
+        digest.update(input, offset, len)
     }
 
     @Test
